requirements.txt
beautifulsoup4==4.4.1
dateutils==0.6.6
docopt==0.6.2
python-dateutil==2.4.2
pytz==2015.7
requests==2.8.1
robobrowser==0.5.3
six==1.10.0
Werkzeug==0.11.2
